remove environment variables related to ConEmu, as they are not passed to the child process
See for example:
http://www.cpantesters.org/cpan/report/8e743c30-6c36-1014-878a-6534a694d5e7

The issue seems to be with unexpected environment variables:
CONEMUARGS, CONEMUANSILOG, CONEMUCONFIG and possibly others that
the test script didn't show. A quick run of googling pointed me to
https://code.google.com/p/conemu-maximus5/

The test script seems to see these environment variables, but somehow they
are not transmitted along to the child process. I suspect this might a
feature of ConEmu itself.

If I'm right, then my only option for now is to remove these variables
from the test data set before running the test.